requirements/base.txt
fastapi==0.110.0
redis==5.0.1
requests==2.31.0
tortoise-orm==0.25.1
jwt==1.4.0
aerich==0.9.1
loguru==0.7.3